+++ src/rendering/RenderUtil.cc
@@ -60,6 +60,7 @@
 #include "ignition/gazebo/components/Scene.hh"
 #include "ignition/gazebo/components/Temperature.hh"
 #include "ignition/gazebo/components/ThermalCamera.hh"
+#include "ignition/gazebo/components/Visibility.hh"
 #include "ignition/gazebo/components/Visual.hh"
 #include "ignition/gazebo/components/World.hh"
 #include "ignition/gazebo/EntityComponentManager.hh"
@@ -168,6 +169,9 @@ class ignition::gazebo::RenderUtilPrivate
   /// \brief A map of entity ids and temperature
   public: std::map<Entity, float> entityTemp;
 
+  /// \brief A map of entity ids and wire boxes
+  public: std::unordered_map<Entity, ignition::rendering::WireBoxPtr> wireBoxes;
+
   /// \brief A map of entity ids and trajectory pose updates.
   public: std::unordered_map<Entity, math::Pose3d> trajectoryPoses;
 
@@ -204,12 +208,10 @@ class ignition::gazebo::RenderUtilPrivate
 
   /// \brief Highlight a node and all its children.
   /// \param[in] _node Node to be highlighted
-  /// TODO(anyone) On future versions, use a bounding box instead
   public: void HighlightNode(const rendering::NodePtr &_node);
 
   /// \brief Restore a highlighted node to normal.
   /// \param[in] _node Node to be restored.
-  /// TODO(anyone) On future versions, use a bounding box instead
   public: void LowlightNode(const rendering::NodePtr &_node);
 };
 
@@ -672,6 +674,7 @@ void RenderUtilPrivate::CreateRenderingEntities(
     _ecm.Each<components::Visual, components::Name, components::Pose,
               components::Geometry,
               components::CastShadows,
+              components::VisibilityFlags,
               components::ParentEntity>(
         [&](const Entity &_entity,
             const components::Visual *,
@@ -679,6 +682,7 @@ void RenderUtilPrivate::CreateRenderingEntities(
             const components::Pose *_pose,
             const components::Geometry *_geom,
             const components::CastShadows *_castShadows,
+            const components::VisibilityFlags *_visibilityFlags,
             const components::ParentEntity *_parent)->bool
         {
           sdf::Visual visual;
@@ -686,6 +690,7 @@ void RenderUtilPrivate::CreateRenderingEntities(
           visual.SetRawPose(_pose->Data());
           visual.SetGeom(_geom->Data());
           visual.SetCastShadows(_castShadows->Data());
+          visual.SetVisibilityFlags(_visibilityFlags->Data());
 
           // Optional components
           auto material = _ecm.Component<components::Material>(_entity);
@@ -842,6 +847,7 @@ void RenderUtilPrivate::CreateRenderingEntities(
     _ecm.EachNew<components::Visual, components::Name, components::Pose,
               components::Geometry,
               components::CastShadows,
+              components::VisibilityFlags,
               components::ParentEntity>(
         [&](const Entity &_entity,
             const components::Visual *,
@@ -849,6 +855,7 @@ void RenderUtilPrivate::CreateRenderingEntities(
             const components::Pose *_pose,
             const components::Geometry *_geom,
             const components::CastShadows *_castShadows,
+            const components::VisibilityFlags *_visibilityFlags,
             const components::ParentEntity *_parent)->bool
         {
           sdf::Visual visual;
@@ -856,6 +863,7 @@ void RenderUtilPrivate::CreateRenderingEntities(
           visual.SetRawPose(_pose->Data());
           visual.SetGeom(_geom->Data());
           visual.SetCastShadows(_castShadows->Data());
+          visual.SetVisibilityFlags(_visibilityFlags->Data());
 
           // Optional components
           auto material = _ecm.Component<components::Material>(_entity);
@@ -1330,45 +1338,47 @@ void RenderUtilPrivate::HighlightNode(const rendering::NodePtr &_node)
 {
   if (!_node)
     return;
-
-  for (auto n = 0u; n < _node->ChildCount(); ++n)
-  {
-    this->HighlightNode(_node->ChildByIndex(n));
-  }
-
   auto vis = std::dynamic_pointer_cast<rendering::Visual>(_node);
-  if (nullptr == vis)
-    return;
-
-  // Visual material
-  auto visMat = vis->Material();
-  if (nullptr != visMat)
+  Entity entityId = kNullEntity;
+  if (vis)
+    entityId = std::get<int>(vis->UserData("gazebo-entity"));
+  // If the entity is not found in the existing map, create a wire box
+  auto wireBoxIt = this->wireBoxes.find(entityId);
+  if (wireBoxIt == this->wireBoxes.end())
   {
-    // If the entity isn't already highlighted, highlight it
-    if (this->originalEmissive.find(vis->Name()) ==
-        this->originalEmissive.end())
+    auto white = this->scene->Material("highlight_material");
+    if (!white)
     {
-      this->originalEmissive[vis->Name()] = visMat->Emissive();
-      visMat->SetEmissive(visMat->Emissive() + math::Color(0.5, 0.5, 0.5));
+      white = this->scene->CreateMaterial("highlight_material");
+      white->SetAmbient(1.0, 1.0, 1.0);
+      white->SetDiffuse(1.0, 1.0, 1.0);
+      white->SetSpecular(1.0, 1.0, 1.0);
+      white->SetEmissive(1.0, 1.0, 1.0);
     }
+
+    ignition::rendering::WireBoxPtr wireBox =
+      this->scene->CreateWireBox();
+    ignition::math::AxisAlignedBox aabb = vis->LocalBoundingBox();
+    wireBox->SetBox(aabb);
+
+    // Create visual and add wire box
+    ignition::rendering::VisualPtr wireBoxVis =
+      this->scene->CreateVisual();
+    wireBoxVis->SetInheritScale(false);
+    wireBoxVis->AddGeometry(wireBox);
+    wireBoxVis->SetMaterial(white, false);
+    vis->AddChild(wireBoxVis);
+
+    // Add wire box to map for setting visibility
+    this->wireBoxes.insert(
+        std::pair<Entity, ignition::rendering::WireBoxPtr>(entityId, wireBox));
   }
-
-  for (auto g = 0u; g < vis->GeometryCount(); ++g)
+  else
   {
-    auto geom = vis->GeometryByIndex(g);
-
-    // Geometry material
-    auto geomMat = geom->Material();
-    if (nullptr == geomMat)
-      continue;
-
-    // If the entity isn't already highlighted, highlight it
-    if (this->originalEmissive.find(geom->Name()) ==
-        this->originalEmissive.end())
-    {
-      this->originalEmissive[geom->Name()] = geomMat->Emissive();
-      geomMat->SetEmissive(geomMat->Emissive() + math::Color(0.5, 0.5, 0.5));
-    }
+    ignition::rendering::WireBoxPtr wireBox = wireBoxIt->second;
+    auto visParent = wireBox->Parent();
+    if (visParent)
+      visParent->SetVisible(true);
   }
 }
 
@@ -1377,53 +1387,16 @@ void RenderUtilPrivate::LowlightNode(const rendering::NodePtr &_node)
 {
   if (!_node)
     return;
-
-  for (auto n = 0u; n < _node->ChildCount(); ++n)
-  {
-    this->LowlightNode(_node->ChildByIndex(n));
-  }
-
   auto vis = std::dynamic_pointer_cast<rendering::Visual>(_node);
-  if (nullptr == vis)
-    return;
-
-  // Visual material
-  auto visMat = vis->Material();
-  if (nullptr != visMat)
+  Entity entityId = kNullEntity;
+  if (vis)
+    entityId = std::get<int>(vis->UserData("gazebo-entity"));
+  if (this->wireBoxes.find(entityId) != this->wireBoxes.end())
   {
-    auto visEmissive = this->originalEmissive.find(vis->Name());
-    if (visEmissive != this->originalEmissive.end())
-    {
-      visMat->SetEmissive(visEmissive->second);
-    }
-    else
-    {
-      ignerr << "Failed to find original material for visual [" << vis->Name()
-             << "]" << std::endl;
-    }
-  }
-
-  for (auto g = 0u; g < vis->GeometryCount(); ++g)
-  {
-    auto geom = vis->GeometryByIndex(g);
-
-    // Geometry material
-    auto geomMat = geom->Material();
-    if (nullptr == geomMat)
-    {
-      ignerr << "Geometry missing material during lowlight." << std::endl;
-      continue;
-    }
-
-    auto geomEmissive = this->originalEmissive.find(geom->Name());
-    if (geomEmissive != this->originalEmissive.end())
-    {
-      geomMat->SetEmissive(geomEmissive->second);
-    }
-    else
-    {
-      ignerr << "Failed to find original material for geometry ["
-             << geom->Name() << "]" << std::endl;
-    }
+    ignition::rendering::WireBoxPtr wireBox =
+      this->wireBoxes[entityId];
+    auto visParent = wireBox->Parent();
+    if (visParent)
+      visParent->SetVisible(false);
   }
 }