+++ Changelog.md
@@ -7,11 +7,17 @@
 1. Actor skeleton animation (auto update mode)
     * [Pull Request 579](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/579)
 
+1. Add support for visual visibility flags and camera visibility mask
+    * [Pull Request 559](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/559)
+
 1. Support <actor><pose> and <actor><plugin>
     * [Pull Request 542](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/542)
 
 1. Depend on ign-rendering4, ign-gui4, ign-sensors4
     * [Pull Request 540](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/540)
+
+1. Axis-Aligned Bounding Boxes
+    * [Pull Request 565](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/565)
 
 ## Ignition Gazebo 3.x